Rename "cache" to "redis" to differentiate with kafka
