+++ .env.sh
@@ -48,7 +48,7 @@ export POSTGRES_HOST=rafiki_db
 export POSTGRES_PORT=5432
 export ADMIN_HOST=rafiki_admin
 export ADMIN_PORT=3000
-export REDIS_HOST=rafiki_cache
+export REDIS_HOST=rafiki_redis
 export REDIS_PORT=6379
 export PREDICTOR_PORT=3003
 export WEB_ADMIN_HOST=rafiki_admin_web
