+++ dockerfiles/worker.Dockerfile
@@ -72,8 +72,10 @@ COPY rafiki/utils/requirements.txt rafiki/utils/requirements.txt
 RUN pip install -r rafiki/utils/requirements.txt
 COPY rafiki/meta_store/requirements.txt rafiki/meta_store/requirements.txt
 RUN pip install -r rafiki/meta_store/requirements.txt
-COPY rafiki/cache/requirements.txt rafiki/cache/requirements.txt
-RUN pip install -r rafiki/cache/requirements.txt
+COPY rafiki/redis/requirements.txt rafiki/redis/requirements.txt
+RUN pip install -r rafiki/redis/requirements.txt
+COPY rafiki/kafka/requirements.txt rafiki/kafka/requirements.txt
+RUN pip install -r rafiki/kafka/requirements.txt
 COPY rafiki/advisor/requirements.txt rafiki/advisor/requirements.txt
 RUN pip install -r rafiki/advisor/requirements.txt
