+++ rafiki/predictor/predictor.py
@@ -26,8 +26,8 @@ import traceback
 
 from rafiki.utils.auth import superadmin_client
 from rafiki.meta_store import MetaStore
-from rafiki.cache import InferenceCache
-from rafiki.cache import RQueue
+from rafiki.redis import InferenceCache as RedisInferenceCache
+from rafiki.kafka import InferenceCache as KafkaInferenceCache
 
 from .constants import Prediction, Query
 from .ensemble import get_ensemble_method
@@ -49,10 +49,10 @@ class Predictor():
         self._inference_job_id = None
 
         self._pull_job_info()
-        self._inference_cache = InferenceCache(self._inference_job_id, 
+        self._redis_cache = RedisInferenceCache(self._inference_job_id, 
                                                 self._redis_host, 
                                                 self._redis_port)
-        self._queue = RQueue()
+        self._kakfa_cache = KafkaInferenceCache()
         logger.info(f'Initialized predictor for inference job "{self._inference_job_id}"')
 
     # Only a single thread should run this
@@ -70,7 +70,7 @@ class Predictor():
 
         # Clear caches for inference job
         try:
-            self._inference_cache.clear_all()
+            self._redis_cache.clear_all()
         except:
             logger.error('Error clearing inference cache:')
             logger.error(traceback.format_exc())
@@ -100,15 +100,15 @@ class Predictor():
         queries = [Query(x) for x in queries]
 
         # Get list of available workers
-        worker_ids = self._inference_cache.get_workers()
+        worker_ids = self._redis_cache.get_workers()
         if len(worker_ids) == 0:
             raise NoWorkersAvailableError()
 
         # For each worker, send queries to worker
         pending_queries = set() # {(query_id, worker_id)}
         for worker_id in worker_ids:
-            self._queue.add_queries_for_worker(worker_id, queries)
-            # self._inference_cache.add_queries_for_worker(worker_id, queries)
+            self._kakfa_cache.add_queries_for_worker(worker_id, queries)
+            # self._redis_cache.add_queries_for_worker(worker_id, queries)
             pending_queries.update([(x.id, worker_id) for x in queries])
 
         # Wait for all predictions to be made
@@ -117,8 +117,8 @@ class Predictor():
             # For every pending query to worker
             for (query_id, worker_id) in list(pending_queries):
                 # Check cache
-                prediction = self._queue.take_prediction_for_worker(worker_id, query_id)
-                # prediction = self._inference_cache.take_prediction_for_worker(worker_id, query_id)
+                prediction = self._kakfa_cache.take_prediction_for_worker(worker_id, query_id)
+                # prediction = self._redis_cache.take_prediction_for_worker(worker_id, query_id)
                 if prediction is None:
                     continue
