+++ rafiki/redis/__init__.py
@@ -20,4 +20,3 @@
 from .param_cache import ParamCache
 from .train_cache import TrainCache
 from .inference_cache import InferenceCache
-from .queue import RQueue
