+++ rafiki/redis/requirements.txt
@@ -1,0 +1,1 @@
+redis==2.10.6
