+++ scripts/start.sh
@@ -29,9 +29,10 @@ bash ./scripts/create_docker_swarm.sh
 bash ./scripts/pull_images.sh || exit 1
 
 # Start whole Rafiki stack
-# Start Zookeeper and Kafka
+# Start Zookeeper, Kafka & Redis
 bash ./scripts/start_zookeeper.sh || exit 1
 bash ./scripts/start_kafka.sh || exit 1
+bash ./scripts/start_redis.sh || exit 1
 # Skip starting & loading DB if DB is already running
 if is_running $POSTGRES_HOST
 then
@@ -40,7 +41,6 @@ else
     bash ./scripts/start_db.sh || exit 1
     bash ./scripts/load_db.sh || exit 1
 fi
-bash ./scripts/start_cache.sh || exit 1
 bash ./scripts/start_admin.sh || exit 1
 bash ./scripts/start_web_admin.sh || exit 1
