+++ test/rafiki/redis/test_param_cache.py
@@ -22,7 +22,7 @@ import numpy as np
 from datetime import datetime, timedelta
 
 from rafiki.advisor import ParamsType
-from rafiki.cache import ParamCache
+from rafiki.redis import ParamCache
 from test.utils import global_setup
 
 class TestParamCache():
